+++ R/check_map_table.R
@@ -15,11 +15,10 @@ check_map_table <- function(map_table) {
   testthat::expect_equal("BP", names(map_table)[4])
   is_too_highs <- map_table$CHR > 95
   chromosome_numbers <- map_table$CHR[is_too_highs]
-
-
   if (any(map_table$CHR > 95)) {
     stop(
-      "PLINK can handle a maximum of 95 chromosomes. \n",
+      "plinkr allows a maximum of 95 chromosomes, as \n",
+      "default PLINK can handle a maximum of 95 chromosomes. \n",
       "(see https://www.cog-genomics.org/plink/1.9/input#chr_set) \n",
       "Found chromosome numbers with values: ",
         paste0(chromosome_numbers, collapse = ", ")