+++ tests/testthat/test-check_map_table.R
@@ -5,7 +5,6 @@ test_that("use", {
 })
 
 test_that("Can handle 95 chromosomes", {
-  # https://www.cog-genomics.org/plink/1.9/input#chr_set
   map_table <- get_test_map_table()
   expect_silent(check_map_table(map_table))
   map_table$CHR[1] <- 96