+++ NEWS.md
@@ -8,6 +8,8 @@ Newest versions at top.
 
  * ?Add vignette with PLINK tutorial
  * Can read `.bed`, `.bim` and `.fam` files
+ * Check that maximally 95 chromosomes are used, as this is a (default)
+   PLINK limit, see https://www.cog-genomics.org/plink/1.9/input#chr_set
 
 ### MINOR IMPROVEMENTS